fixed final scene

+++ bonappetit/Assets/GetFinalScore.cs
@@ -7,7 +7,7 @@ using TMPro;
 
 public class GetFinalScore : MonoBehaviourPunCallbacks
 {
-    private int finalScore;
+    private float finalScore;
     private TMP_Text tm;
     private PhotonView photonView;
     private GameObject textGo;
@@ -27,14 +27,14 @@ public class GetFinalScore : MonoBehaviourPunCallbacks
     void Update()
     {
         //tm.text = finalScore.ToString();
-        int score = (int) PhotonNetwork.CurrentRoom.CustomProperties["score"];
+        float score = (float) PhotonNetwork.CurrentRoom.CustomProperties["score"];
         tm.text = score.ToString();
     }
 
     public override void OnRoomPropertiesUpdate(ExitGames.Client.Photon.Hashtable propertiesThatChanged)
     {
         if (propertiesThatChanged.ContainsKey("score")) {
-            finalScore = (int) propertiesThatChanged["score"];
+            finalScore = (float) propertiesThatChanged["score"];
         }
     }
 }
